requirements-test.txt
# Testing dependencies for TalkingPhoto AI MVP

# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.3.1
pytest-timeout==2.2.0
pytest-env==1.1.0

# Test data and factories
factory-boy==3.3.0
faker==20.1.0
freezegun==1.2.2

# HTTP and API testing
httpx==0.25.2
responses==0.23.3
requests-mock==1.11.0

# Database testing
pytest-postgresql==5.0.0
pytest-redis==3.0.2

# Performance and load testing
locust==2.17.0
memory-profiler==0.61.0
pytest-benchmark==4.0.0

# Mock and testing utilities
unittest-mock==1.0.1
testfixtures==7.2.0
betamax==0.8.1

# Coverage and reporting
coverage[toml]==7.3.2
pytest-html==4.1.1
pytest-json-report==1.5.0

# Linting and code quality (for CI)
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
bandit==1.7.5

# Documentation testing
pytest-doctestplus==1.0.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Specific testing for Streamlit
streamlit==1.28.1

# AI service testing mocks
openai==1.3.5
anthropic==0.8.0

# Image and video processing for tests
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# Database and storage testing
SQLAlchemy==2.0.23
redis==5.0.1
boto3==1.34.0  # For S3 testing

# Security testing
cryptography==41.0.8
passlib==1.7.4

# Logging and monitoring
structlog==23.2.0

# Utilities
click==8.1.7
python-dateutil==2.8.2